+++ tests/test_reference.py
@@ -345,6 +345,19 @@ def test_generate_report():
           - SUCCESS: computed the correct median
     """).strip()
 
+    # check group
+    res = ref.run(footprint, group="median")
+    report = generate_report(res)
+    assert report == dedent("""\
+        REFERENCE: foo
+        GROUP: median
+        SATISFIED: True
+        MESSAGES:
+          - SUCCESS: Sorted the sample correctly
+          - SUCCESS: Computed the size of the sample
+          - SUCCESS: computed the correct median
+    """).strip()
+
     # check empty messages
     ref = ReferenceImplementation("foo", [Value(footprint.get_value(0).value)])
     res = ref.run(footprint)